+++ qiskit/transpiler/passes/routing/algorithms/util.py
